Feature: Support for Volume Targets
This allows you to override the keyboard media keys, to allow the volume buttons to control a different audio device (or session) than the system default. For instance, the normal volume could control the default audio device, but Shift+Volume controls could control the volume of your headphones, and Control+Volume could control the audio of just your browser.

This system is perhaps not as clean as it could be, but it is reasonably robust about allowing you to set a target of an application on a specific audio device and properly handling even when the application is quit and restarted.